requirements.txt
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
nltk==3.9.2
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.1
regex==2026.1.15
textblob==0.19.0
tqdm==4.67.3
Werkzeug==3.1.5
